+++ pyinaturalist/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa: F401, F403
-__version__ = '0.14.1'
-DEFAULT_USER_AGENT = f'pyinaturalist/{__version__}'
+__version__ = '0.15.0'
+DEFAULT_USER_AGENT: str = f'pyinaturalist/{__version__}'
 user_agent = DEFAULT_USER_AGENT
 
 # Ignore ImportErrors if this is imported outside a virtualenv